// src/app/models/outra-pericia.ts
export class OutrasPericias {
    codigo: any;
    usuario: any;
    //
    tipo: any;
    nome_tipo: any;
    //
    idade: any;
    peso: any;
    estatura: any;
    descricao: any;
    //
    tipo_local: any;
    nome_tipo_local: any;
    //
    examen_requisitado: any;
    nome_examen_requisitado: any;
    descricao_do_local: any;
    quesitaces: any;
    numguia: any;
    numocur: any;
    rascunho: any;
    tipo_guia: string;

    constructor() {

    }
}